+++ lib/db/src/schema/index.ts
@@ -146,3 +146,4 @@ export * from "./aiExecution";
 export * from "./aiModelCapabilities";
 export * from "./aiEvaluation";
 export * from "./aiClinicalConfig";
+export * from "./aiInterop";